Tell the indexes how a hash emptied, not just that it did
A key that ends up with no fields left is either a refusal or a document with
nothing in it, and which one it is depends on the command. HDEL of the last
field sends the indexes to read a key that is not there and is counted in
hash_indexing_failures. A deadline that took the last field writes the document
one more time with nothing in it and is counted as nothing, so it spends a
number where HDEL spends none. HSETEX with a deadline that has already passed is
two pieces of news from one command, so the number moves twice and the value
never reaches the index at all.

All three are measured against 8.10.1 and none of them is guessable. The signal
coming out of the hash group is now a four state Change rather than a bool.

records_per_doc_avg is also divided in single precision and printed in double,
which is what a real server does and shows up in the digits.

+++ crates/yo-resp/src/dispatch/search.rs
@@ -1234,7 +1234,18 @@ fn info<'a>(reg: &mut Registry, args: Args<'a>, out: &mut Out) -> Answer<'a> {
     // build does not keep yet. All four are a division by nought on an empty
     // index, and a real server answers `nan` there for the same reason rather
     // than as a placeholder.
-    number(out, "records_per_doc_avg", records as f64 / docs as f64);
+    //
+    // The division is done in single precision and printed in double, which is
+    // what a real server does and is visible: seventeen records over three
+    // documents comes out as 5.666666507720947 and not 5.666666666666667. The
+    // counters are doubles by the time they reach the reply and the arithmetic
+    // behind them is not, so a client comparing the two builds sees the same
+    // digits.
+    number(
+        out,
+        "records_per_doc_avg",
+        f64::from(records as f32 / docs as f32),
+    );
     number(out, "bytes_per_record_avg", f64::NAN);
     number(out, "offsets_per_term_avg", f64::NAN);
     number(out, "offset_bits_per_record_avg", f64::NAN);
